feat(reservation): Theme 의존성 추가에 따른 쿼리 변경 및 매퍼 로직 변경
