+++ src/main/java/roomescape/domain/time/ReservationTimeRepository.java
@@ -1,12 +1,9 @@
 package roomescape.domain.time;
 
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.sql.Time;
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 
@@ -44,13 +41,5 @@ public class ReservationTimeRepository {
     }, generatedKeyHolder);
     return new ReservationTime(generatedKeyHolder.getKeyAs(Long.class),
         newReservationTime.startAt());
-  }
-
-  private class ReservationTimeRowMapper implements RowMapper<ReservationTime> {
-
-    @Override
-    public ReservationTime mapRow(ResultSet rs, int rowNum) throws SQLException {
-      return new ReservationTime(rs.getLong("id"), rs.getTime("start_at").toLocalTime());
-    }
   }
 }
