+++ src/main/java/roomescape/domain/time/ReservationTimeRowMapper.java
@@ -1,0 +1,13 @@
+package roomescape.domain.time;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import org.springframework.jdbc.core.RowMapper;
+
+public class ReservationTimeRowMapper implements RowMapper<ReservationTime> {
+
+  @Override
+  public ReservationTime mapRow(ResultSet rs, int rowNum) throws SQLException {
+    return new ReservationTime(rs.getLong("id"), rs.getTime("start_at").toLocalTime());
+  }
+}
